refactor: improve architecture
