+++ src/Product/Show/Details/Gallery/styled.js
